GPU functions for total exclusion energy
Previous code stores exclusion and self-energy of molecules (assume rigid) and does not calculate it every time (saves time). 
Now, an option to calculate it during the simulation is provided on the GPU as a __global__ kernel function. 

NOTE: This is only applied to the TOTAL energy calculation, not available in MC moves... Something to keep in mind!