framework component fix for translation/rotation move
This fixed a horrible mistake from before. 
NCrossAtom was used to replace "NHostAtom" in the previous codes. 
It was detected when the code is used for a system where NGuestAtom > NHostAtom (and then for framework component being translated, the code breaks and result in energy drift).
Now (hopefully) fixed. Need to consider making the change for framework component that can do CBMC (might be interest for future users).